feat: switch from got to ofetch
