+++ .eslintrc
@@ -4,5 +4,7 @@
     "node": true
   },
   "extends": ["@ctrl/eslint-config"],
-  "rules": {}
+  "rules": {
+    "no-bitwise": "off"
+  }
 }
