+++ package.json
@@ -28,24 +28,23 @@
     "test:ci": "vitest run --coverage --reporter=default --reporter=junit --outputFile=./junit.xml"
   },
   "dependencies": {
-    "@ctrl/magnet-link": "^3.1.1",
-    "@ctrl/shared-torrent": "^4.3.1",
-    "@ctrl/torrent-file": "^2.0.2",
-    "@ctrl/url-join": "^2.0.2",
-    "formdata-node": "^5.0.0",
-    "got": "^12.5.0",
-    "tough-cookie": "^4.1.2"
+    "@ctrl/magnet-link": "^3.1.2",
+    "@ctrl/shared-torrent": "^5.0.0",
+    "@ctrl/torrent-file": "^3.0.0",
+    "node-fetch-native": "^1.4.0",
+    "ofetch": "^1.3.3",
+    "tough-cookie": "^4.1.3",
+    "ufo": "^1.3.0"
   },
   "devDependencies": {
-    "@ctrl/eslint-config": "3.5.0",
-    "@sindresorhus/tsconfig": "3.0.1",
-    "@types/node": "18.7.18",
-    "@types/tough-cookie": "4.0.2",
-    "@vitest/coverage-c8": "0.23.4",
-    "c8": "7.12.0",
-    "typedoc": "0.23.15",
-    "typescript": "4.8.3",
-    "vitest": "0.23.4"
+    "@ctrl/eslint-config": "4.0.6",
+    "@sindresorhus/tsconfig": "4.0.0",
+    "@types/node": "20.6.0",
+    "@types/tough-cookie": "4.0.3",
+    "@vitest/coverage-v8": "0.34.4",
+    "typedoc": "0.25.1",
+    "typescript": "5.2.2",
+    "vitest": "0.34.4"
   },
   "publishConfig": {
     "access": "public"
@@ -56,6 +55,6 @@
     ]
   },
   "engines": {
-    "node": ">=14.16"
+    "node": ">=18"
   }
 }
