+++ vitest.config.ts
@@ -3,5 +3,9 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     threads: false,
+    coverage: {
+      reporter: ['text', 'json', 'html'],
+      provider: 'v8',
+    },
   },
 });
